+++ testfiles/attrs.py
@@ -1,0 +1,3 @@
+foo.bar.bang
+
+x = x.y.z