+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * Added type `attr` (eg. `attr#foo` will match `a.foo.bar` and `a.bar.foo`)
+* Added support for wildcards
 
 ### Bug fixed